requirements.txt
fastapi
uvicorn
python-dotenv
supabase
google-genai
pydantic
pydantic-settings
